The board is now scrollable.
To prevent an AssertionError from Pyglet, I had to check, before deleting a VertexList, that its attribute domain.allocator.starts evaluates to True. domain.allocator.starts is actually a List--I think it's a list of indices into the VertexList?--and if it's empty (therefore, False), I think that means you have a vertex list with nothing to draw, and therefore, nothing to delete.