rename iter_places, iter_place_bones -> iter_places_hosted, iter_place_bones_hosted
I think I need to redo the interface so that when you are adding things to the world, you are adding them to the charsheet, and vice versa. You can't add things to the world without adding them to a character, so why should it be possible to keep them off of the charsheets?