No special tables for things or portals, only their locations and stats.
Neither things nor portals really exist without locations anyway.

Besides, this makes it possible for more than one character to contain
the same thing or portal. That's neat. Well, it's messy, but in an
intentional kind of way.

It's possible, now, for Things and Portals to exist with only a host,
not actually being part of a character the way I had them.
I mean, they need to be IN a character, because characters have graphs,
and you gotta be in a graph to be in any type of game space--
although it seems to be possible to give stats to things that don't
exist. Those will show up on character sheets although the thing in
question won't show anywhere.

Anyway, yeah, got rid of some tables and integrity constraints I
didn't think I'd need.