remove fire_time_travel_triggers
This was a function to poll for changes to stats, idea being to detect
when a stat changed because it was scheduled to, even though __getitem__
would never be called in that instance.

The function was unbearably slow. I may rewrite it.
But I'm trying to find another way to achieve the same effect.