Get rid of everything to do with cross-process callbacks
Proxies will hereafter poll the LiSE core when you ask them to.

Maybe I'll implement a way to subscribe to changes in LiSE someday
but it seems more logical to do that from the proxies, and use
a thread to listen for LiSE to tell you "stuff's changed, poll
if you want to know more".