Move do_game_start and module installation into EngineProxy
This makes it convenient to do both things after the subprocess is
spawned but before caching anything from it.

I can't think of a reason to do anything else at that moment so
I'm hardcoding it. Maybe I'll put a hook in for
'run this every time the game starts'? But wouldn't that make
more sense to do in ELiDE, in the parent process?