Mostly get the character switcher working at speed
It still isn't switching the stats column to the new
char correctly. There seem to be several issues with
that; sometimes there's a mismatch between which character's
initially loaded, and which is displayed in the stats
column.

I didn't do anything to make new boards when you make
new characters in ELiDE.