Rewrite loaders to enable automatic branching
Previously, it was possible for ELiDE to automatically make
a new branch when the user contradicted history,
but this would only happen if they contradicted history for
a *specific* entity; making new history for something that
had been to that point unchanged, but when there were already
changes recorded to other entities later in the same branch,
was allowed, when in hindsight it should not have been.

Now HistoryError is appropriately raised whenever history
is contradicted at all, but the error isn't handled
appropriately yet.