Speed up time travel in ELiDE
Basically, I'm not returning deltas to ELiDE anymore. When it requests time travel, it gets a whole keyframe. At least, in the simplest case.

LiSE may still use deltas internally to build the requested keyframe. When it does that, it's from the most recent keyframe prior, rather than whatever time the user is looking at. If LiSE is taking regular keyframes as requested by its `keyframe_interval` option, this will result in far less comparison for the delta function to perform.

I observe this approach to work well in practice.